Revert "Extract SystemTableProvider from Hive"
This reverts commit 4f1fa43e61b14af494e237f01b1da4120b3a5fd0.
