+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/SystemTableProvider.java
@@ -11,9 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.trino.plugin.base.connector;
+package io.trino.plugin.hive;
 
-import io.trino.spi.connector.ConnectorMetadata;
 import io.trino.spi.connector.ConnectorSession;
 import io.trino.spi.connector.SchemaTableName;
 import io.trino.spi.connector.SystemTable;
@@ -24,5 +23,5 @@ public interface SystemTableProvider
 {
     Optional<SchemaTableName> getSourceTableName(SchemaTableName table);
 
-    Optional<SystemTable> getSystemTable(ConnectorMetadata metadata, ConnectorSession session, SchemaTableName tableName);
+    Optional<SystemTable> getSystemTable(HiveMetadata metadata, ConnectorSession session, SchemaTableName tableName);
 }
